# docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: taxbook-pro API
  version: 1.0.0
  description: |
    REST API for taxbook-pro.

    ## Authentication
    All endpoints (except where noted) require a valid Bearer token in the Authorization header.

    ## Rate Limiting
    - Standard: 100 requests/minute
    - Authenticated: 1000 requests/minute

    ## Pagination
    List endpoints support cursor-based pagination via `cursor` and `limit` query parameters.
  contact:
    name: API Support
    email: api@example.com
  license:
    name: Proprietary
    url: https://example.com/license
servers:
  - url: https://api.example.com/v1
    description: Production
  - url: https://api.staging.example.com/v1
    description: Staging
  - url: http://localhost:3000/api
    description: Local Development

# ============================================================
# SECURITY SCHEMES
# ============================================================
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from authentication endpoint.
        Include in Authorization header: `Bearer <token>`

# ============================================================
# REUSABLE SCHEMAS
# ============================================================
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required:
            - type
            - message
          properties:
            type:
              type: string
              enum:
                - bad_request
                - unauthorized
                - forbidden
                - not_found
                - conflict
                - rate_limited
                - internal_error
            message:
              type: string
              example: The requested resource was not found
            details:
              type: array
              items:
                $ref: '#/components/schemas/ValidationError'
            requestId:
              type: string
              format: uuid
              example: 550e8400-e29b-41d4-a716-446655440000
    ValidationError:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
          example: email
        message:
          type: string
          example: Invalid email format
        code:
          type: string
          example: invalid_format

    # --- Pagination ---
...
    Profile:
      type: object
      required:
        - id
        - id
        - userId
        - email
        - name
        - timezone
        - subscriptionTier
        - maxDailyAppointments
        - maxDailyAppointmentsTaxSeason
        - createdAt
        - updatedAt
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the 
          example: 550e8400-e29b-41d4-a716-446655440000
        id:
          type: string
          example: 'Example id'
        userId:
          type: string
          example: 'Example user_id'
          description: References auth.users(id)
        email:
          type: string
          example: 'Example email'
        name:
          type: string
          example: 'Example name'
        firmName:
          type: string
          example: 'Example firm_name'
          description: Name of the tax preparation firm
          nullable: true
        licenseNumber:
          type: string
          example: 'Example license_number'
          description: CPA or tax preparer license number
          nullable: true
        timezone:
          type: string
          example: 'Example timezone'
          description: User&#x27;s timezone for scheduling
        subscriptionTier:
          type: string
          example: 'Example subscription_tier'
        bookingSlug:
          type: string
          example: 'Example booking_slug'
          description: Unique slug for public booking page
          nullable: true
        taxSeasonStart:
          type: string
          example: 'tax_season_start'
          description: When tax season starts (for capacity limits)
          nullable: true
        taxSeasonEnd:
          type: string
          example: 'tax_season_end'
          description: When tax season ends
          nullable: true
        maxDailyAppointments:
          type: number
          example: 100
          description: Maximum appointments per day during normal times
        maxDailyAppointmentsTaxSeason:
          type: number
          example: 100
          description: Maximum appointments per day during tax season
        createdAt:
          type: string
          example: 'created_at'
        updatedAt:
          type: string
          example: 'updated_at'
        createdAt:
          type: string
          format: date-time
          description: When the Profile was created
          example: '2024-01-15T10:30:00Z'
        updatedAt:
          type: string
          format: date-time
          description: When the Profile was last updated
          example: '2024-01-15T14:45:00Z'
    CreateProfileInput:
      type: object
      required:
        - userId
        - email
        - name
        - timezone
        - subscriptionTier
        - maxDailyAppointments
        - maxDailyAppointmentsTaxSeason
      properties:
        userId:
          type: string
          example: 'Example user_id'
          description: References auth.users(id)
        email:
          type: string
          example: 'Example email'
        name:
          type: string
          example: 'Example name'
        firmName:
          type: string
          example: 'Example firm_name'
          description: Name of the tax preparation firm
        licenseNumber:
          type: string
          example: 'Example license_number'
          description: CPA or tax preparer license number
        timezone:
          type: string
          example: 'Example timezone'
          description: User&#x27;s timezone for scheduling
        subscriptionTier:
          type: string
          example: 'Example subscription_tier'
        bookingSlug:
          type: string
          example: 'Example booking_slug'
          description: Unique slug for public booking page
        taxSeasonStart:
          type: string
          example: 'tax_season_start'
          description: When tax season starts (for capacity limits)
        taxSeasonEnd:
          type: string
          example: 'tax_season_end'
          description: When tax season ends
        maxDailyAppointments:
          type: number
          example: 100
          description: Maximum appointments per day during normal times
        maxDailyAppointmentsTaxSeason:
          type: number
          example: 100
          description: Maximum appointments per day during tax season
    UpdateProfileInput:
      type: object
      properties:
        userId:
          type: string
          example: 'Example user_id'
          description: References auth.users(id)
        email:
          type: string
          example: 'Example email'
        name:
          type: string
          example: 'Example name'
        firmName:
          type: string
          example: 'Example firm_name'
          description: Name of the tax preparation firm
        licenseNumber:
          type: string
          example: 'Example license_number'
          description: CPA or tax preparer license number
        timezone:
          type: string
          example: 'Example timezone'
          description: User&#x27;s timezone for scheduling
        subscriptionTier:
          type: string
          example: 'Example subscription_tier'
        bookingSlug:
          type: string
          example: 'Example booking_slug'
          description: Unique slug for public booking page
        taxSeasonStart:
          type: string
          example: 'tax_season_start'
          description: When tax season starts (for capacity limits)
        taxSeasonEnd:
          type: string
          example: 'tax_season_end'
          description: When tax season ends
        maxDailyAppointments:
          type: number
          example: 100
          description: Maximum appointments per day during normal times
        maxDailyAppointmentsTaxSeason:
          type: number
          example: 100
          description: Maximum appointments per day during tax season
    ProfileListResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          required:
            - items
            - total
            - page
            - limit
            - hasMore
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Profile'
            total:
              type: integer
              example: 100
            page:
              type: integer
              example: 1
            limit:
              type: integer
              example: 20
            hasMore:
              type: boolean
              example: true
            nextCursor:
              type: string
              example: eyJpZCI6IjEyMyJ9
    ProfileResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Profile'
    ProfileDeleteResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          required:
            - deleted
          properties:
            deleted:
              type: boolean
              example: true

    # --- Client ---
...
  parameters:
    PathId:
      name: id
      in: path
      required: true
      description: Resource UUID
      schema:
        type: string
        format: uuid
      example: 550e8400-e29b-41d4-a716-446655440000
    QueryPage:
      name: page
      in: query
      required: false
      description: Page number for pagination (1-indexed)
      schema:
        type: integer
        minimum: 1
        default: 1
      example: 1
    QueryLimit:
      name: limit
      in: query
      required: false
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      example: 20
    QueryCursor:
      name: cursor
      in: query
      required: false
      description: Cursor for pagination (use nextCursor from previous response)
      schema:
        type: string
      example: eyJpZCI6IjEyMyJ9
    QueryOrderBy:
      name: orderBy
      in: query
      required: false
      description: Field to order results by
      schema:
        type: string
        default: createdAt
      example: createdAt
    QueryOrderDir:
      name: orderDir
      in: query
      required: false
      description: Order direction
      schema:
        type: string
        enum: [asc, desc]
        default: desc
      example: desc

# ============================================================
# REUSABLE RESPONSES
# ============================================================
  responses:
    BadRequest:
      description: Bad Request - Invalid input or validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              type: bad_request
              message: Validation failed
              details:
                - field: email
                  message: Invalid email format
                  code: invalid_format
              requestId: 550e8400-e29b-41d4-a716-446655440000
    Unauthorized:
      description: Unauthorized - Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              type: unauthorized
              message: Authentication required
              requestId: 550e8400-e29b-41d4-a716-446655440000
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              type: forbidden
              message: You do not have permission to access this resource
              requestId: 550e8400-e29b-41d4-a716-446655440000
    NotFound:
      description: Not Found - Resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              type: not_found
              message: The requested resource was not found
              requestId: 550e8400-e29b-41d4-a716-446655440000
    InternalError:
      description: Internal Server Error - Unexpected error occurred
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              type: internal_error
              message: An unexpected error occurred
              requestId: 550e8400-e29b-41d4-a716-446655440000

# ============================================================
# API PATHS
# ============================================================
paths:
...
  /profile:
    get:
      tags:
        - Profile
      summary: List Profile
      description: |
        Retrieve a paginated list of Profile records.
        Supports filtering, sorting, and cursor-based pagination.
      operationId: listProfile
      parameters:
        - $ref: '#/components/parameters/QueryPage'
        - $ref: '#/components/parameters/QueryLimit'
        - $ref: '#/components/parameters/QueryCursor'
        - $ref: '#/components/parameters/QueryOrderBy'
        - $ref: '#/components/parameters/QueryOrderDir'
        - name: email
          in: query
          required: false
          description: Filter by email
          schema:
            type: string
        - name: name
          in: query
          required: false
          description: Filter by name
          schema:
            type: string
      responses:
        '200':
          description: Successful response with list of Profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags:
        - Profile
      summary: Create Profile
      description: Create a new Profile record.
      operationId: createProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProfileInput'
      responses:
        '201':
          description: Successfully created Profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /profile/{id}:
    get:
      tags:
        - Profile
      summary: Get Profile by ID
      description: Retrieve a single Profile record by its unique identifier.
      operationId: getProfile
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: Successful response with Profile details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    patch:
      tags:
        - Profile
      summary: Update Profile
      description: |
        Update an existing Profile record.
        Only provided fields will be updated (partial update).
      operationId: updateProfile
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileInput'
      responses:
        '200':
          description: Successfully updated Profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags:
        - Profile
      summary: Delete Profile
      description: |
        Delete a Profile record by its unique identifier.
      operationId: deleteProfile
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: Successfully deleted Profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileDeleteResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # --- Client Endpoints ---
...